backend/requirements.txt
# ========================================
# AI-Powered Job Portal - Python Dependencies
# ========================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Machine Learning & NLP
scikit-learn==1.3.2
numpy==1.26.2

# Document Processing
pypdf==3.17.1
python-docx==1.1.0

# Email Support
aiosmtplib==3.0.1
email-validator==2.1.0
jinja2==3.1.2

# Validation & Settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Additional utilities
python-dateutil==2.8.2